pyproject.toml
[project]
name = "api-teste-mercafacil"
version = "0.1.0"
description = ""
authors = [
    {name = "Vinicius José",email = "vinijosenog@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "toml (>=0.10.2,<0.11.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
